+++ NEWS.md
@@ -47,6 +47,7 @@
 * Commas linter handles missing arguments calls properly (#145)
 * Add `function_left_parentheses_linter` to check that there is no space between
   a function name and its left parentheses (#204, @jrnold).
+* Implement `summary.lints()` (#260, #262, @wlandau).
 
 # lintr 1.0.1 #
 * bugfix to work with knitr 1.16.7