+++ NEWS.md
@@ -1,3 +1,6 @@
+# lintr 1.0.2 #
+* Fix tests to work with upcoming testthat release.
+
 # lintr 1.0.1 #
 * bugfix to work with knitr 1.16.7
 * `expect_lint_free()` now is always skipped on CRAN. This is necessary because